Clean up song library screen
